+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.1.7 (2024-09-24)
-1. fix `utils.code_inline` gzip compress `mtime` to 0
+1. fix `utils.code_inline` gzip compress `mtime` to 1, to ensure the same compressed data
 2. fix `utils.FileDict.save`
 3. add `ipc.QueueManager` based on BaseManager, add JSON listener
 4. add `compress`, `ensure_dir` arg to `funtools.SizedTimedRotatingFileHandler`